all u8 to f64 bar dmx send

+++ src/json_storage.rs
@@ -2,7 +2,7 @@ use serde_json::Error;
 use std::fs::File;
 use std::io::{BufReader, BufWriter, Write};
 
-pub fn write_to_file(lighting_records: &Vec<(String, Vec<u8>)>) -> std::io::Result<()> {
+pub fn write_to_file(lighting_records: &Vec<(String, Vec<f64>)>) -> std::io::Result<()> {
     // SO 69449293
     let file = File::create("lighting_records.json")?;
     let mut writer = BufWriter::new(file);
@@ -11,7 +11,7 @@ pub fn write_to_file(lighting_records: &Vec<(String, Vec<u8>)>) -> std::io::Resu
     Ok(())
 }
 
-pub fn read_from_file() -> Result<Vec<(String, Vec<u8>)>, Error> {
+pub fn read_from_file() -> Result<Vec<(String, Vec<f64>)>, Error> {
     // will panic if no file
     let file = File::open("lighting_records.json").expect("File not found or corrupt");
     let reader = BufReader::new(file);
